requirements.txt
absl-py==2.1.0
astroid==2.3.3
astunparse==1.6.3
atomicwrites==1.4.1
attrs==24.2.0
charset-normalizer==3.4.0
click==8.1.8
coverage==7.2.7
cycler==0.11.0
Flask==2.2.5
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.38.0
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.62.3
h5py==3.8.0
idna==3.10
importlib-metadata==6.7.0
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.5
joblib==1.2.0
keras==2.11.0
Keras-Applications==1.0.8
keras-vggface==0.6
kiwisolver==1.4.5
lazy-object-proxy==1.4.3
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mccabe==0.6.1
more-itertools==9.1.0
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.2.0.32
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==24.0
pathlib2==2.3.4
Pillow==9.5.0
pluggy==0.13.1
protobuf==3.19.6
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydot==2.0.0
pydotplus==2.0.2
pylint==2.4.4
pyparsing==3.1.4
pytest==5.1.3
pytest-cov==2.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.10.3
python-json-logger==0.1.11
pywin32==303
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.17.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
typed-ast==1.4.3
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.11.2
zipp==3.15.0
